requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.9
colorama==0.4.4
crispy-tailwind==0.3.0
Django==3.1.7
django-crispy-forms==1.11.1
django-environ==0.4.5
django-redis==5.0.0
et-xmlfile==1.1.0
idna==2.10
Jinja2==3.0.3
json-tricks==3.15.5
MarkupSafe==2.0.1
numpy==1.20.3
openai==0.11.4
openpyxl==3.0.7
pandas==1.2.4
pandas-stubs==1.2.0.39
plotly==4.14.3
predictor-api-client==1.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
tqdm==4.62.3
urllib3==1.26.4
